[lv2] proper include statement for lv2.h
With lv2 version 1.18.6, the file /usr/include/lv2.h no longer exists.
It is now in /usr/include/lv2/core/lv2.h
However, this file does not exists with previous version of lv2.
It appears that, at least for version 1.18.4, a copy of lv2.h is also in
/usr/include/lv2/lv2plug.in/ns/lv2core/ Hopefully, using this path keeps
some backward compatibility